Change type of Point::x and Point::y (uint32 -> int32)

+++ src/UpdraftPkg/Updraft/Graphics/Circle.cpp
@@ -8,7 +8,7 @@ void Circle::draw(const Color color) const
   {
     for(int32 x = pos.x - r; x < pos.x + r; x++)
     {
-      if(Math::Pow(x - (int32)pos.x, 2) + Math::Pow(y - (int32)pos.y, 2) <= Math::Pow(r, 2)) Point(x, y).draw(color);
+      if(Math::Pow(x - pos.x, 2) + Math::Pow(y - pos.y, 2) <= Math::Pow(r, 2)) Point(x, y).draw(color);
     }
   }
 }
